requirements.txt
streamlit==1.28.1
google-generativeai==0.8.0
chromadb==0.4.24
pandas==2.1.4
python-dotenv==1.0.0
numpy==1.26.2
scikit-learn==1.3.2